# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartReserve — Smart Demand-Aware Offer Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers/{offerId}/slots:
    get:
      operationId: listSlots
      tags: [slots]
      summary: List slots for an offer
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
    post:
      operationId: createSlot
      tags: [slots]
      summary: Create a slot for an offer
      security:
        - bearerAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slot"
  /slots/{id}:
    patch:
      operationId: updateSlot
      tags: [slots]
      summary: Update a slot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slot"
    delete:
      operationId: deleteSlot
      tags: [slots]
      summary: Delete a slot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # BOOKINGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Slot:
      type: object
      required: [id, offerId, slotDate, startTime, endTime, capacity, bookedCount, availableCount, status]
      properties:
        id:
          type: integer
        offerId:
          type: integer
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        bookedCount:
          type: integer
        availableCount:
          type: integer
        status:
          type: string
        occupancyPercent:
          type: number
        demandLevel:
          type: string
    SlotInput:
      type: object
      required: [slotDate, startTime, endTime, capacity]
      properties:
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
    SlotUpdate:
      type: object
      properties:
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        status:
          type: string

    # BOOKING
...